// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  CS: 'CS',
  TECHNICIAN: 'TECHNICIAN',
  STAFF: 'STAFF'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const AuthProvider = {
  EMAIL: 'EMAIL',
  GOOGLE: 'GOOGLE',
  BOTH: 'BOTH'
} as const

export type AuthProvider = (typeof AuthProvider)[keyof typeof AuthProvider]


export const IncomeSource = {
  CUSTOMER_PAYMENT: 'CUSTOMER_PAYMENT',
  MANUAL: 'MANUAL'
} as const

export type IncomeSource = (typeof IncomeSource)[keyof typeof IncomeSource]